Rename to photobooth.py and convert tabs to spaces
